refactor(*): Add reset support in Accessor.

+++ magneton-modules/magneton-module-safedog/src/main/java/org/magneton/module/safedog/access/AccessTimeCalculator.java
@@ -15,6 +15,6 @@ public interface AccessTimeCalculator {
 	 * @param accessConfig 访问配置
 	 * @return 锁定时间，单位毫秒
 	 */
-	long calculate(String name, int wrongs, AccessConfig accessConfig);
+	long calculate(String name, long wrongs, AccessConfig accessConfig);
 
 }
